Separate placeholder tests from main TXTDirect tests
